Update proxy_requests.py
Modified the script to use BeautifulSoup for parsing, which results in more proxies in the sockets list. Also implemented the sockets as a Singleton, to avoid loading the same list of proxies for every instance. 
The ProxyRequests class now requires the user to manually call the acquire_sockets method (at least once) to load the list of usable servers.